added UrlManager
