+++ repo/src/main/java/com/parashift/onlyoffice/scripts/DownloadAs.java
@@ -1,6 +1,7 @@
 package com.parashift.onlyoffice.scripts;
 
 import com.parashift.onlyoffice.util.ConvertManager;
+import com.parashift.onlyoffice.util.UrlManager;
 import com.parashift.onlyoffice.util.Util;
 import org.alfresco.model.ContentModel;
 import org.alfresco.repo.download.DownloadModel;
@@ -74,6 +75,9 @@ public class DownloadAs extends AbstractWebScript {
     @Autowired
     MessageService mesService;
 
+    @Autowired
+    UrlManager urlManager;
+
     @Override
     public void execute(WebScriptRequest request, WebScriptResponse response) throws IOException {
         try {
@@ -98,9 +102,9 @@ public class DownloadAs extends AbstractWebScript {
                 if (currentExt.equals(outputType)) {
                     contentURL = getDownloadAPIUrl(node, docTitle);
                 } else {
-                    String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, util.getContentUrl(node), region);
+                    String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, urlManager.getContentUrl(node), region);
                     docTitle = docTitle.substring(0, docTitle.lastIndexOf(".") + 1) + outputType;
-                    URL url = new URL(util.replaceDocEditorURLToInternal(downloadUrl));
+                    URL url = new URL(urlManager.replaceDocEditorURLToInternal(downloadUrl));
                     HttpURLConnection connection = (HttpURLConnection) url.openConnection();
                     try (InputStream inputStream = connection.getInputStream()) {
                         contentURL = createDownloadNode(docTitle, mimetypeService.getMimetype(outputType), inputStream, connection.getContentLength(), 1);
@@ -138,9 +142,9 @@ public class DownloadAs extends AbstractWebScript {
                                 totalSize = totalSize + IOUtils.copyLarge(inputStream, out);
                             }
                         } else {
-                            String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, util.getContentUrl(node), region);
+                            String downloadUrl = converterService.convert(util.getKey(node), currentExt, outputType, urlManager.getContentUrl(node), region);
                             docTitle = docTitle.substring(0, docTitle.lastIndexOf(".") + 1) + outputType;
-                            URL url = new URL(util.replaceDocEditorURLToInternal(downloadUrl));
+                            URL url = new URL(urlManager.replaceDocEditorURLToInternal(downloadUrl));
                             HttpURLConnection connection = (HttpURLConnection) url.openConnection();
                             try (InputStream inputStream = connection.getInputStream()) {
                                 out.putArchiveEntry(new ZipArchiveEntry(docTitle));
