+++ repo/src/main/java/com/parashift/onlyoffice/scripts/EditorApi.java
@@ -2,6 +2,7 @@ package com.parashift.onlyoffice.scripts;
 
 import com.parashift.onlyoffice.util.ConvertManager;
 import com.parashift.onlyoffice.util.JwtManager;
+import com.parashift.onlyoffice.util.UrlManager;
 import com.parashift.onlyoffice.util.Util;
 
 import org.alfresco.model.ContentModel;
@@ -67,6 +68,9 @@ public class EditorApi extends AbstractWebScript {
     @Autowired
     MessageService mesService;
 
+    @Autowired
+    UrlManager urlManager;
+
     private Logger logger = LoggerFactory.getLogger(this.getClass());
 
     @Override
@@ -111,7 +115,7 @@ public class EditorApi extends AbstractWebScript {
                         data.put("c", requestData.get("command"));
                     }
                     data.put("fileType", fileType);
-                    data.put("url", util.getContentUrl(node));
+                    data.put("url", urlManager.getContentUrl(node));
                     if (jwtManager.jwtEnabled()) {
                         try {
                             data.put("token", jwtManager.createToken(data));
@@ -159,7 +163,7 @@ public class EditorApi extends AbstractWebScript {
                 }
 
                 try {
-                    String downloadUrl = converterService.convert(util.getKey(node), fileType, "docxf", util.getContentUrl(node), mesService.getLocale().toLanguageTag());
+                    String downloadUrl = converterService.convert(util.getKey(node), fileType, "docxf", urlManager.getContentUrl(node), mesService.getLocale().toLanguageTag());
                     docTitle = docTitle.substring(0, docTitle.lastIndexOf("."));
                     String newNode = createNode(folderNode, docTitle, "docxf", downloadUrl);
                     data.put("nodeRef", newNode);
@@ -203,7 +207,7 @@ public class EditorApi extends AbstractWebScript {
     }
 
     private String createNode(NodeRef folderNode, String title, String ext, String url) throws IOException {
-        url = util.replaceDocEditorURLToInternal(url);
+        url = urlManager.replaceDocEditorURLToInternal(url);
         String fileName = util.getCorrectName(folderNode, title, ext);
 
         NodeRef nodeRef = nodeService.createNode(
