+++ repo/src/main/java/com/parashift/onlyoffice/util/UrlManager.java
@@ -1,0 +1,168 @@
+package com.parashift.onlyoffice.util;
+
+import org.alfresco.repo.admin.SysAdminParams;
+import org.alfresco.repo.imap.ImapService;
+import org.alfresco.service.cmr.coci.CheckOutCheckInService;
+import org.alfresco.service.cmr.repository.NodeRef;
+import org.alfresco.service.cmr.repository.NodeService;
+import org.alfresco.service.cmr.repository.StoreRef;
+import org.alfresco.service.cmr.security.AuthenticationService;
+import org.alfresco.util.UrlUtil;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.extensions.surf.util.URLEncoder;
+import org.springframework.stereotype.Service;
+
+import java.util.Arrays;
+import java.util.Date;
+import java.util.List;
+
+/*
+    Copyright (c) Ascensio System SIA 2022. All rights reserved.
+    http://www.onlyoffice.com
+*/
+@Service
+public class UrlManager {
+
+    @Autowired
+    @Qualifier("checkOutCheckInService")
+    CheckOutCheckInService cociService;
+
+    @Autowired
+    AuthenticationService authenticationService;
+
+    @Autowired
+    SysAdminParams sysAdminParams;
+
+    @Autowired
+    NodeService nodeService;
+
+    @Autowired
+    ConfigManager configManager;
+
+    @Autowired
+    ImapService imapService;
+
+    public String getCreateNewUrl(NodeRef nodeRef, String docType){
+        String folderNodeRef = this.nodeService.getPrimaryParent(nodeRef).getParentRef().toString();
+        String docMime = "application/vnd.openxmlformats-officedocument.wordprocessingml.document";
+        switch (docType) {
+            case "cell": {
+                docMime = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet";
+                break;
+            }
+            case "slide": {
+                docMime = "application/vnd.openxmlformats-officedocument.presentationml.presentation";
+                break;
+            }
+        }
+        return getShareUrl() + "page/onlyoffice-edit?parentNodeRef=" + folderNodeRef + "&new=" + docMime;
+    }
+
+    public String getFavoriteUrl(NodeRef nodeRef){
+        return "parashift/onlyoffice/editor-api/favorite?nodeRef=" + nodeRef.toString();
+    }
+
+    public String getHistoryInfoUrl(NodeRef nodeRef) {
+        return "parashift/onlyoffice/history/info?nodeRef=" + nodeRef.toString();
+    }
+
+    public String getHistoryDataUrl(NodeRef nodeRef) {
+        return "parashift/onlyoffice/history/data?nodeRef=" + nodeRef.toString();
+    }
+
+    public String getHistoryDiffUrl(NodeRef nodeRef) {
+        return getAlfrescoUrl() + "s/parashift/onlyoffice/download/diff?nodeRef=" + nodeRef.toString() + "&alf_ticket=" + authenticationService.getCurrentTicket();
+    }
+
+    public String getContentUrl(NodeRef nodeRef) {
+        return  getAlfrescoUrl() + "s/parashift/onlyoffice/download/file?nodeRef=" + nodeRef.toString() + "&alf_ticket=" + authenticationService.getCurrentTicket();
+    }
+
+    public String getCallbackUrl(NodeRef nodeRef) {
+        String hash = null;
+        if (cociService.isCheckedOut(nodeRef)) {
+            hash = (String) nodeService.getProperty(cociService.getWorkingCopy(nodeRef), Util.EditingHashAspect);
+        }
+
+        return getAlfrescoUrl() + "s/parashift/onlyoffice/callback?nodeRef=" + nodeRef.toString() + "&cb_key=" + hash;
+    }
+
+    public String getTestConversionUrl() {
+        return getAlfrescoUrl() + "s/parashift/onlyoffice/convertertest?alf_ticket=" + authenticationService.getCurrentTicket();
+    }
+
+    public String getEditorUrl() {
+        return configManager.demoActive() ? configManager.getDemo("url") : (String) configManager.getOrDefault("url", "http://127.0.0.1/");
+    }
+
+    public String getBackUrl(NodeRef nodeRef){
+        String url = imapService.getContentFolderUrl(nodeRef);
+
+        if (url.contains("?filter=path|")) {
+            List<String> urlParts = Arrays.asList(url.split("\\|"));
+            url = urlParts.get(0) + URLEncoder.encodeUriComponent("|" + urlParts.get(1));
+        }
+
+        return url;
+    }
+
+    public String getEditorInnerUrl() {
+        String url = (String) configManager.getOrDefault("innerurl", "");
+        if (url.isEmpty() || configManager.demoActive()) {
+            return getEditorUrl();
+        } else {
+            return url;
+        }
+    }
+
+    public String getEmbeddedSaveUrl(String sharedId, String docTitle) {
+        StringBuilder embeddedSaveUrl = new StringBuilder(8);
+        embeddedSaveUrl.append(UrlUtil.getShareUrl(sysAdminParams));
+        embeddedSaveUrl.append("/proxy/alfresco-noauth/api/internal/shared/node/");
+        embeddedSaveUrl.append(sharedId);
+        embeddedSaveUrl.append("/content/");
+        embeddedSaveUrl.append(URLEncoder.encodeUriComponent(docTitle));
+        embeddedSaveUrl.append("?c=force");
+        embeddedSaveUrl.append("&noCache=" + new Date().getTime());
+        embeddedSaveUrl.append("&a=true");
+
+        return embeddedSaveUrl.toString();
+    }
+
+    public String getEmbeddedSaveUrl(NodeRef nodeRef, String docTitle) {
+        StringBuilder embeddedSaveUrl = new StringBuilder(7);
+        StoreRef storeRef = nodeRef.getStoreRef();
+        embeddedSaveUrl.append(UrlUtil.getShareUrl(sysAdminParams));
+        embeddedSaveUrl.append("/proxy/alfresco/slingshot/node/content");
+        embeddedSaveUrl.append("/" + storeRef.getProtocol());
+        embeddedSaveUrl.append("/" + storeRef.getIdentifier());
+        embeddedSaveUrl.append("/" + nodeRef.getId());
+        embeddedSaveUrl.append("/" + URLEncoder.encodeUriComponent(docTitle));
+        embeddedSaveUrl.append("?a=true");
+
+        return embeddedSaveUrl.toString();
+    }
+
+    private String getAlfrescoUrl() {
+        String alfUrl = (String) configManager.getOrDefault("alfurl", "");
+        if (alfUrl.isEmpty()) {
+            return UrlUtil.getAlfrescoUrl(sysAdminParams) + "/";
+        } else {
+            return alfUrl + "alfresco/";
+        }
+    }
+
+    public String getShareUrl(){
+        return UrlUtil.getShareUrl(sysAdminParams) + "/";
+    }
+
+    public String replaceDocEditorURLToInternal(String url) {
+        String innerDocEditorUrl = getEditorInnerUrl();
+        String publicDocEditorUrl = getEditorUrl();
+        if (!publicDocEditorUrl.equals(innerDocEditorUrl) && !configManager.demoActive()) {
+            url = url.replace(publicDocEditorUrl, innerDocEditorUrl);
+        }
+        return url;
+    }
+}
